refactor estimation.py into submods
